# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Agri-Fi API",
    "description": "REST API for the Agri-Fi agricultural trade finance platform. Farmers list produce, traders create deals, investors fund them via Stellar escrow.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    },
    {
      "url": "https://api.agri-fi.example.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      }
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Amara Diallo"
                  },
                  "email": {
                    "type": "string",
                    "example": "amara@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "securePass1"
                  },
                  "role": {
                    "enum": ["farmer", "trader", "investor", "company_admin"],
                    "example": "trader"
                  },
                  "country": {
                    "type": "string",
                    "example": "KE"
                  }
                },
                "required": ["name", "email", "password", "role", "country"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Authenticate and receive a JWT",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "amara@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "securePass1"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns access_token JWT"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/investments/my-investments": {
      "get": {
        "operationId": "InvestmentsController_getMyInvestments",
        "summary": "List the authenticated investor's investments",
        "tags": [
          "investments"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "example": 1,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "example": 20,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of investments"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only investors can access this endpoint"
          }
        }
      }
    },
...
  }
}